requirements.txt
qdrant-client==1.12.1
langchain==0.3.14
langchain-community==0.3.12
openai==1.59.7
python-dotenv==1.0.0
pdfplumber==0.5.28
gradio==5.12.0
tiktoken==0.8.0
nltk==3.9.1